update with new facedb, bilinear params

+++ examples/anime_viewer/facedb/facedb.h
@@ -37,23 +37,19 @@ class FaceDB
     static const std::vector<int>           gLandmarks73;
 public:
 #ifdef USE_75_45
-
     static const int LengthIdentity = 75;
 #ifdef PARAMETER_FACS
     static const int LengthExpression = 47;
 #else
     static const int LengthExpression = 45;
 #endif
-
 #else
-
     static const int LengthIdentity = 50;
 #ifdef PARAMETER_FACS
     static const int LengthExpression = 47;
 #else
     static const int LengthExpression = 25;
 #endif
-
 #endif
 
     static double MAX_ALLOWED_WEIGHT_RANGE;
@@ -61,7 +57,7 @@ public:
     static int                                  NumDimIden()    { return gTensorShape[1];     }
     static int                                  NumDimExpr()    { return gTensorShape[2];     }
     static int                                  NumVertices()   { return gTensorShape[0] / 3; }
-    static int                                  NumTriangles()  { return gTriangles.size();   }
+    static int                                  NumTriangles()  { return (int)gTriangles.size();   }
     static const Tensor3 &                      CoreTensor()    { return gCoreTensor;         }
 
     static const MatrixRM &                     IdenUT()        { return gIdenUT; }
